# apps/web/src/shared/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fleet Maintenance Decision API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/inventory": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Inventory",
        "operationId": "inventory_api_inventory_get",
        "parameters": [
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventoryPartResponse"
                  },
                  "title": "Response Inventory Api Inventory Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/inventory/{part_id}/adjustments": {
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Adjustment",
        "operationId": "adjustment_api_inventory__part_id__adjustments_post",
        "parameters": [
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Part Id"
            }
          },
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockAdjustment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Adjustment Api Inventory  Part Id  Adjustments Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/inventory/arrivals": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Arrivals",
        "operationId": "arrivals_api_inventory_arrivals_get",
        "parameters": [
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArrivalResponse"
                  },
                  "title": "Response Arrivals Api Inventory Arrivals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Expected Delivery",
        "operationId": "expected_delivery_api_inventory_arrivals_post",
        "parameters": [
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArrivalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArrivalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/inventory/arrivals/{arrival_id}/outcome": {
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Delivery Outcome",
        "operationId": "delivery_outcome_api_inventory_arrivals__arrival_id__outcome_post",
        "parameters": [
          {
            "name": "arrival_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Arrival Id"
            }
          },
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArrivalOutcome"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArrivalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArrivalOutcome": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "receive",
              "cancel",
              "quarantine",
              "reject"
            ],
            "title": "Action"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "action",
          "expected_version",
          "reason"
        ],
        "title": "ArrivalOutcome"
      },
      "ArrivalRequest": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.-]+$",
            "title": "Id"
          },
          "part_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Part Id"
          },
          "quantity": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 1.0,
            "title": "Quantity"
          },
          "arrival_slot": {
            "type": "integer",
            "maximum": 13.0,
            "minimum": 0.0,
            "title": "Arrival Slot"
          },
          "expected_part_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Part Version"
          },
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "part_id",
          "quantity",
          "arrival_slot",
          "expected_part_version",
          "reason"
        ],
        "title": "ArrivalRequest"
      },
      "ArrivalResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "part_id": {
            "type": "string",
            "title": "Part Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "arrival_slot": {
            "type": "integer",
            "title": "Arrival Slot"
          },
          "status": {
            "type": "string",
            "enum": [
              "expected",
              "received",
              "cancelled",
              "quarantined",
              "rejected"
            ],
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "received_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At"
          },
          "provenance": {
            "type": "string",
            "const": "synthetic",
            "title": "Provenance"
          },
          "slot_duration_hours": {
            "type": "integer",
            "const": 8,
            "title": "Slot Duration Hours"
          }
        },
        "type": "object",
        "required": [
          "id",
          "part_id",
          "quantity",
          "arrival_slot",
          "status",
          "version",
          "reason",
          "created_at",
          "received_at",
          "provenance",
          "slot_duration_hours"
        ],
        "title": "ArrivalResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InventoryPartResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "on_hand": {
            "type": "integer",
            "minimum": 0.0,
            "title": "On Hand"
          },
          "reserved": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Reserved"
          },
          "lead_time_slots": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Lead Time Slots"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "provenance": {
            "type": "string",
            "title": "Provenance"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "on_hand",
          "reserved",
          "lead_time_slots",
          "version",
          "provenance"
        ],
        "title": "InventoryPartResponse"
      },
...
      "StockAdjustment": {
        "properties": {
          "delta_free_stock": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": -1000000.0,
            "title": "Delta Free Stock"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason"
          },
          "command_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Command Id"
          }
        },
        "type": "object",
        "required": [
          "delta_free_stock",
          "expected_version",
          "reason",
          "command_id"
        ],
        "title": "StockAdjustment"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}